working version of regular schedule and optimal allocation, new file names and reorganized code. deleting now old files including helper_functions, processing_functions, back_star_functions, and reporting_functions. New files history, io, weather and access
