+++ kpfcc/onsky.py
@@ -24,9 +24,8 @@ import ttp.model as model
 
 # path2modules = os.path.dirname(os.path.abspath(__file__))
 # sys.path.append(path2modules)
-import kpfcc.backup_star_functions as bsf
-import kpfcc.helper_functions as hf
-import kpfcc.processing_functions as pf
+import kpfcc.weather as wt
+import kpfcc.io as io
 
 named_colors = ['blue', 'red', 'green', 'gold', 'maroon', 'gray', 'orange', 'magenta', 'purple']
 
@@ -69,7 +68,7 @@ def run_ttp(manager):
 
     # Prepare relevant files
     round_two_requests = np.loadtxt(output_path + 'Round2_Requests.txt', dtype=str)
-    send_to_ttp = pf.prepare_for_ttp(manager.request_frame, the_schedule[day_in_semester],
+    send_to_ttp = prepare_for_ttp(manager.request_frame, the_schedule[day_in_semester],
                                         round_two_requests)
     filename = output_path + 'Selected_' + str(manager.current_day) + ".txt"
     send_to_ttp.to_csv(filename, index=False)
@@ -85,7 +84,7 @@ def run_ttp(manager):
     plotting.plot_path_2D(solution,outputdir=observers_path)
     plotting.nightPlan(solution.plotly, manager.current_day, outputdir=observers_path)
     obs_and_times = pd.read_csv(observers_path + 'ObserveOrder_' + str(manager.current_day) + ".txt")
-    pf.write_starlist(manager.request_frame, solution.plotly, observation_start_time, solution.extras,
+    io.write_starlist(manager.request_frame, solution.plotly, observation_start_time, solution.extras,
                         round_two_requests, str(manager.current_day), observers_path)
     print("The optimal path through the sky for the selected stars is found. Clear skies!")
 
@@ -116,12 +115,12 @@ def produce_bright_backups(backup_file, backup_observability_file, observatory,
     backup_starlist = pd.read_csv(backup_file)
     backup_observability_frame = pd.read_csv(backup_observability_file)
 
-    backups = bsf.get_stars_for_tonight(backup_starlist, backup_observability_frame,
+    backups = wt.get_stars_for_tonight(backup_starlist, backup_observability_frame,
             current_day[5:], minimum_up_time=4.0)
     backups.drop(columns='index', inplace=True)
 
     filename_backups = backups_path + '/Backups_' + str(current_day) + ".txt"
-    backups_full = bsf.get_times_worth(backups, total_time+1)
+    backups_full = wt.get_times_worth(backups, total_time+1)
     backups_full.to_csv(filename_backups, index=False)
     targlist_backups = formatting.theTTP(filename_backups)
 
@@ -134,6 +133,56 @@ def produce_bright_backups(backup_file, backup_observability_file, observatory,
     plotting.plot_path_2D(solution_b, outputdir = backups_path)
     plotting.nightPlan(solution_b.plotly, current_day, outputdir = backups_path)
     obs_and_times_b = pd.read_csv(backups_path + 'ObserveOrder_' + current_day + ".txt")
-    pf.write_starlist(backup_starlist, solution_b.plotly, observation_start_time,
+    io.write_starlist(backup_starlist, solution_b.plotly, observation_start_time,
                         solution_b.extras, [], str(current_day), backups_path)
     print("Bright backups script created.")
+
+def prepare_for_ttp(request_frame, night_plan, round_two_targets):
+    """
+    Prepare tonight's scheduled stars for their run through the TTP (separate software package)
+
+    Args:
+        request_sheet (dataframe): the csv of PI requests
+        night_plan (array): the n'th row of combined_semester_schedule array
+        round_two_targets (array): a 1D list of the stars that were added in the bonus round
+
+    Returns:
+        to_ttp (dataframe): the data on the stars to be observed tonight, formatted in the way that
+                            the TTP software requires as an input
+    """
+    ignore = ['*', 'W', '', '*X', 'X']
+    selected_stars = []
+    for i, item in enumerate(night_plan):
+        if night_plan[i] not in ignore and night_plan[i][:4] != "RM___":
+            selected_stars.append(night_plan[i])
+            ignore.append(night_plan[i])
+
+    starnames = []
+    ras = []
+    decs = []
+    exposure_times = []
+    exposures_per_visit = []
+    visits_in_night = []
+    cadences = []
+    priorities = []
+    for j, item in enumerate(selected_stars):
+        idx = request_frame.index[request_frame['Starname']==str(selected_stars[j])][0]
+        starnames.append(str(request_frame['Starname'][idx]))
+        ras.append(request_frame['RA'][idx])
+        decs.append(request_frame['Dec'][idx])
+        exposure_times.append(int(request_frame['Nominal Exposure Time [s]'][idx]))
+        exposures_per_visit.append(int(request_frame['# of Exposures per Visit'][idx]))
+        visits_in_night.append(int(request_frame['# Visits per Night'][idx]))
+        cadences.append(int(request_frame['Minimum Intra-Night Cadence'][idx]))
+        # higher numbers are higher priorities, filler targets get low priority
+        if str(selected_stars[j]) in round_two_targets:
+            prior = 1
+        else:
+            prior = 10
+        priorities.append(prior)
+    to_ttp = pd.DataFrame({"Starname":starnames,"RA":ras,"Dec":decs,
+                          "Exposure Time":exposure_times,
+                          "Exposures Per Visit":exposures_per_visit,
+                          "Visits In Night":visits_in_night, "Intra_Night_Cadence":cadences,
+                          "Priority":priorities})
+    return to_ttp
