+++ kpfcc/schedule.py
@@ -20,13 +20,11 @@ import pandas as pd
 import gurobipy as gp
 from gurobipy import GRB
 
-import kpfcc.helper_functions as hf
-import kpfcc.twilight_functions as tw
-import kpfcc.reporting_functions as rf
-import kpfcc.processing_functions as pf
-import kpfcc.mapping_functions as mf
-import kpfcc.constraint_functions as cf
-import kpfcc.admin_functions as af
+import kpfcc.io as io
+import kpfcc.access as ac
+import kpfcc.management as mn
+import kpfcc.maps as mp
+import kpfcc.constraints as cf
 
 # import line_profiler
 # @profile
@@ -44,16 +42,12 @@ def run_kpfcc(manager):
     round_info = 'Round1'
 
     print("Reading inputs and prepping files.")
-    default_access_maps = mf.construct_access_dict(manager.accessibilities_file, manager.requests_frame)
-    custom_access_maps = mf.construct_custom_map_dict(manager.special_map_file)
-    zero_out_names = mf.construct_zero_out_arr(manager.zero_out_file)
-    twilight_map_remaining_2D, available_slots_in_each_night = tw.construct_twilight_map(manager)
-    nonqueue_map_file_slots_ints = mf.construct_nonqueue_arr(manager)
+    twilight_map_remaining_2D, available_slots_in_each_night = ac.construct_twilight_map(manager)
 
     # When running a normal schedule, include the observatory's allocation map
     if manager.run_optimal_allocation == False:
         weather_diff_remaining, allocation_map_1D, allocation_map_2D, weathered_map = \
-                                    mf.prepare_allocation_map(manager, available_slots_in_each_night)
+                                    mp.prepare_allocation_map(manager, available_slots_in_each_night)
     # When running the optimal allocation, all dates are possible except for those specifically blacked out
     else:
         # Weather arrays are zeros since no weather losses are modeled
@@ -66,16 +60,11 @@ def run_kpfcc(manager):
     manager.weather_diff_remaining = weather_diff_remaining
 
     print("Computing intersection of all maps. ")
-    available_indices_for_request = mf.produce_ultimate_map(manager,
-                                                            allocation_map_1D.flatten(),
-                                                            twilight_map_remaining_2D.flatten(),
-                                                            default_access_maps,
-                                                            custom_access_maps,
-                                                            zero_out_names,
-                                                            nonqueue_map_file_slots_ints)
+    available_indices_for_request = mp.produce_ultimate_map(manager, allocation_map_1D.flatten(),
+                                                            twilight_map_remaining_2D.flatten())
 
     print("Building Gorubi model.")
-    Aframe, Aset, schedulable_requests = hf.define_slot_index_frame(manager, available_indices_for_request)
+    Aframe, Aset, schedulable_requests = cf.define_slot_index_frame(manager, available_indices_for_request)
     model = cf.GorubiModel(manager, Aset, Aframe, schedulable_requests)
     model.constraint_build_theta()
     model.constraint_one_request_per_slot()
@@ -90,9 +79,9 @@ def run_kpfcc(manager):
         model.constraint_forbidden_quarter_patterns()
         model.constraint_cannot_observe_if_not_allocated(twilight_map_remaining_2D)
         if os.path.exists(manager.blackout_file):
-            constraint_enforce_restricted_nights(manager.blackout_file, limit=0)
+            constraint_enforce_restricted_nights(limit=0)
         if os.path.exists(manager.whiteout_file):
-            constraint_enforce_restricted_nights(manager.whiteout_file, limit=1)
+            constraint_enforce_restricted_nights(limit=1)
         if manager.include_aesthetic:
             model.constraint_max_consecutive_onsky()
             model.constraint_minimum_consecutive_offsky()
@@ -113,18 +102,18 @@ def run_kpfcc(manager):
                 allocation_schedule_1d.append(0)
         allocation_schedule = np.reshape(allocation_schedule_1d, (manager.n_nights_in_semester, manager.n_quarters_in_night))
         holder = np.zeros(np.shape(allocation_schedule))
-        allocation_map_1D, allocation_map_2D, weathered_map = mf.build_allocation_map(allocation_schedule, holder, available_slots_in_each_night, manager.n_slots_in_night)
-        mf.convert_allocation_array_to_binary(allocation_schedule, manager.all_dates_array, manager.output_directory + "optimal_allocation_binary_schedule.txt")
+        allocation_map_1D, allocation_map_2D, weathered_map = mp.build_allocation_map(allocation_schedule, holder, available_slots_in_each_night, manager.n_slots_in_night)
+        mp.convert_allocation_array_to_binary(allocation_schedule, manager.all_dates_array, manager.output_directory + "optimal_allocation_binary_schedule.txt")
         manager.allocation_all = allocation_map_1D
-        
+
     print("Building human readable schedule.")
-    combined_semester_schedule_available = hf.write_available_human_readable(manager,
+    combined_semester_schedule_available = io.write_available_human_readable(manager,
                                         twilight_map_remaining_2D, allocation_map_2D, weathered_map)
-    combined_semester_schedule_stars = hf.write_stars_schedule_human_readable(
+    combined_semester_schedule_stars = io.write_stars_schedule_human_readable(
                                                     combined_semester_schedule_available, model.Yrds,
                                                     manager, round_info)
-    rf.build_fullness_report(combined_semester_schedule_stars, allocation_map_2D, manager, round_info)
-    rf.write_out_results(manager, model, round_info, start_the_clock)
+    io.build_fullness_report(combined_semester_schedule_stars, allocation_map_2D, manager, round_info)
+    io.write_out_results(manager, model, round_info, start_the_clock)
 
     if manager.run_round_two:
         print("Beginning Round 2 Scheduling.")
@@ -135,11 +124,11 @@ def run_kpfcc(manager):
 
         print("Total Time to finish Round 2 solver: ", np.round(time.time()-start_the_clock,3))
 
-        combined_semester_schedule_stars = hf.write_stars_schedule_human_readable(
+        combined_semester_schedule_stars = io.write_stars_schedule_human_readable(
                 combined_semester_schedule_available, model.Yrds, manager, round_info)
-        rf.build_fullness_report(combined_semester_schedule_stars, allocation_map_2D, manager, round_info)
-        rf.get_gap_filler_targets(manager)
-        rf.write_out_results(manager, model, round_info, start_the_clock)
+        io.build_fullness_report(combined_semester_schedule_stars, allocation_map_2D, manager, round_info)
+        mn.get_gap_filler_targets(manager)
+        io.write_out_results(manager, model, round_info, start_the_clock)
 
     else:
         print("Not running Round 2. Duplicating Raw Schedule as dummy file.")
