+++ kpfcc/twilight_functions.py
@@ -13,156 +13,3 @@ from astropy.time import TimeDelta
 import astroplan as apl
 
 import kpfcc.mapping_functions as mf
-
-def construct_twilight_map(manager):
-    """
-    Compute the number of slots per night available, based on strictly twilight times
-
-    Args:
-        current_day (str): today's date, format YYYY-MM-DD
-        twilight_frame (dataframe): the pandas dataframe containing twilight information
-        slot_size (int): the size of a single slot, in minutes
-        all_dates_dict (dictionary): keys are calendar days, values are day of the semester
-        n_slots_in_night (int): the number of slots in single night, regardless of twilight
-        n_nights_in_semester (int): the number of nigths remaining in the semester
-
-    Returns:
-        twilight_map_remaining_flat (array): a 1D array where elements are 1 or 0 based if outside or inside of twilight time
-        twilight_map_remaining_2D (array): a 2D version of the same information
-    """
-    print("Determine available slots in each night.")
-    # available_slots_in_each_night is a 1D matrix of length nights n
-    # This is not a gorubi variable, but a regular python variable
-    # Each element will hold an integer which represents the number of slots are available in each
-    # quarter of a given night, after accounting for non-observable times due to day/twilight.
-    available_slots_in_each_night = []
-    for date in list(manager.all_dates_array):
-        slots_tonight = determine_twilight_edge(date, manager.twilight_frame, manager.slot_size)
-        available_slots_in_each_night.append(slots_tonight)
-    twilight_map_all = np.array(mf.build_twilight_map(available_slots_in_each_night,
-                                manager.n_slots_in_night, invert=False))
-    twilight_map_remaining = twilight_map_all[manager.all_dates_dict[manager.current_day]:]
-    twilight_map_remaining_1D = twilight_map_remaining.copy().flatten()
-    twilight_map_remaining_2D = np.reshape(twilight_map_remaining_1D, (manager.n_nights_in_semester, manager.n_slots_in_night))
-    return twilight_map_remaining_2D, available_slots_in_each_night
-
-def convert_slot_to_quarter(twilight_map_remaining_2D_d):
-    '''
-    Determine the slot numbers within the night that breaks the night into "equal" length quarters
-    Take extra precaution when the total number of slots between twilight times is not easily
-    divisable by 4.
-    '''
-
-    n_available_slots_in_quarter_tonight = int(np.sum(twilight_map_remaining_2D_d)/4)
-    extra_slots = np.sum(twilight_map_remaining_2D_d)%4
-    first_slot = np.argmax(twilight_map_remaining_2D_d)
-
-    if extra_slots == 0:
-        # when night is naturally divided into 4, accept as is
-        split_1st2nd = first_slot + n_available_slots_in_quarter_tonight
-        split_2nd3rd = split_1st2nd + n_available_slots_in_quarter_tonight
-        split_3rd4th = split_2nd3rd + n_available_slots_in_quarter_tonight
-    elif extra_slots == 1 or extra_slots == 2:
-        # when night has 1 extra slot, we place it into the 1st quarter
-        # when night has 2 extra slots, we place one into 1st, and one into 4th
-        split_1st2nd = first_slot + n_available_slots_in_quarter_tonight + 1
-        split_2nd3rd = split_1st2nd + n_available_slots_in_quarter_tonight
-        split_3rd4th = split_2nd3rd + n_available_slots_in_quarter_tonight
-    elif extra_slots == 3:
-        # when night has 3 extra slots, we place one into 1st, one into 3rd, and one into 4th
-        split_1st2nd = first_slot + n_available_slots_in_quarter_tonight + 1
-        split_2nd3rd = split_1st2nd + n_available_slots_in_quarter_tonight
-        split_3rd4th = split_2nd3rd + n_available_slots_in_quarter_tonight + 1
-    return split_1st2nd, split_2nd3rd, split_3rd4th
-
-def generate_twilight_times(all_dates_array):
-    """generate_twilight_times
-
-    Precompute a dataframe of the morning/evening twilight times for each day in the semester
-
-    Args:
-        all_dates_array (list): the calendar dates of the semester.
-                                Format: YYYY-MM-DD.
-    Returns:
-        twilight_frame (dataframe): the precomputed twilight times
-    """
-    keck = apl.Observer.at_site('W. M. Keck Observatory')
-
-    twilight_frame = pd.DataFrame({'time_utc':all_dates_array})
-    eighteen_deg_evening = []
-    twelve_deg_evening = []
-    six_deg_evening = []
-    eighteen_deg_morning = []
-    twelve_deg_morning = []
-    six_deg_morning = []
-    for day in twilight_frame.index.strftime(date_format='%Y-%m-%d').tolist():
-        as_day = Time(day,format='iso',scale='utc')
-        eighteen_deg_evening.append(keck.twilight_evening_astronomical(as_day,which='next'))
-        twelve_deg_evening.append(keck.twilight_evening_nautical(as_day,which='next'))
-        six_deg_evening.append(keck.twilight_evening_civil(as_day,which='next'))
-        eighteen_deg_morning.append(keck.twilight_morning_astronomical(as_day,which='next'))
-        twelve_deg_morning.append(keck.twilight_morning_nautical(as_day,which='next'))
-        six_deg_morning.append(keck.twilight_morning_civil(as_day,which='next'))
-
-    twilight_frame['18_evening'] = eighteen_deg_evening
-    twilight_frame['12_evening'] = twelve_deg_evening
-    twilight_frame['6_evening'] = six_deg_evening
-    twilight_frame['18_morning'] = eighteen_deg_morning
-    twilight_frame['12_morning'] = twelve_deg_morning
-    twilight_frame['6_morning'] = six_deg_morning
-
-    return twilight_frame
-
-def get_available_slots_per_night(all_dates_array, twilight_frame, slot_size=5, start_time='03:30',
-                                  n_hours_in_night=14):
-    """get_available_slots_per_night
-
-    Generates a 1D list of length equal to the number of nights in the semester where each element
-    is the integer number of slots that are available to the scheduler in that night
-
-    Args:
-        all_dates_array (list): the calendar dates of the semester.
-                            Format: YYYY-MM-DD.
-        twilight_frame (dataframe): the precomputed twilight times
-    Returns:
-        None
-    """
-    available_slots_per_night = []
-    for date in range(all_dates_array):
-        edge = determine_twilight_edge(date, twilight_frame, slot_size, start_time,
-                                       n_hours_in_night)
-        available_slots_per_night.append(edge)
-    return available_slots_per_night
-
-def determine_twilight_edge(date, twilight_frame, slot_size, start_time='03:30',
-                            n_hours_in_night=14):
-    """determine_twilight_edge
-
-    Determine how many slots within each night actually take place during night time.
-
-    Args:
-        - date (string): the calendar date of the semester. Format: YYYY-MM-DD.
-        - twilight_frame (dataframe): the dataframe containing pre-computed
-        - slot_size (int): the size of the a slot in minutes
-        - start_time (str): the time representing the first slot of the night, format HH:MM
-        - n_hours_in_night (int): number of hours in a night
-    Returns:
-        available_slots_in_night (int): the number of slots in night that are night time
-    """
-    ind = twilight_frame.index[twilight_frame['time_utc'] == date].tolist()
-    twilight_evening_time_isot = twilight_frame.loc[ind,'12_evening'].values[0]
-    twilight_evening_time_bjd = Time(twilight_evening_time_isot,format='jd')
-    start_timestamp = date + "T" + start_time
-    start_timestamp_time = Time(start_timestamp)
-    n_slots = int((n_hours_in_night*60)/slot_size)
-
-    bin_edges = []
-    bin_edges.append(str(start_timestamp_time))
-    for i in range(n_slots):
-        start_timestamp_time += TimeDelta(slot_size/(24*60),format='jd')
-        if twilight_evening_time_bjd < str(start_timestamp_time):
-            edge = i
-            break
-
-    available_slots_in_night = n_slots - 2*edge
-    return available_slots_in_night
